[ruby] Silence frozen string mutated deprecation warnings
From Ruby 3.4 literal strings are frozen by default. Attempting to
mutate them emits a deprecation warning, with the intention of raising
an error in a future version of Ruby.

For these tests we can enable the pragma `# frozen_string_literal:
false` to disable this behaviour and avoid unnecessary string copying in
the affected files.