Exit with an error if PROTOBUF_CONFIG is set incorrectly
Now that there is more than one value that PROTOBUF_CONFIG can have we
should start to be a little more careful about checking its value.

This commit raises an error during build if PROTOBUF_CONFIG has been
explicitly set to value that is not in the allowed config list
(currently 'dbg' and 'asan').

If it's been set to an empty string, the build will continue with the
default (non-debug) parameters, but will print a warning message, to
help the user determine whether this is intended behaviour or not.