Ruby: Optionally warn when arena chains grow large
This PR implements an optional warning, enabled when Ruby is run with
-W2 verbose mode, that warns when an arena fusion occurs that causes an
arena list to grow beyond a certain user defined size defined by
PROTOBUF_ARENA_WARNING_THRESHOLD.

The following warning will be emitted when a fusion list crosses the
threshold

    warning: Large arena growth detected: 373072 bytes, 1001 arenas.

This is an attempt to mitigate problems arising from fusing arenas
attached to objects with mismatched lifetimes as seen in the issues
below.

https://github.com/protocolbuffers/protobuf/issues/19498
https://github.com/protocolbuffers/protobuf/issues/10106

Ideally we'd detect when an arena was being fused to an existing arena
list with a longer lifetime, but this isn't possible with the
information we have about the Ruby object. So this felt like the
simplest, least intrusive alternative.

We're hoping that this will be useful for sense-checking protobuf client
code in CI, or for assessing the cause for memory issues we see in
production.

Given the following test script

```
require 'bundler/inline'

gemfile do
  source 'https://rubygems.org'
  gemspec path: __dir__
  gem 'bigdecimal'
end

require_relative 'tests/basic_test_proto2_pb'

msg = BasicTestProto2::TestMessage.new(
  optional_string: "x" * 10_000,  # 10KB
  optional_bytes: ("y" * 10_000).b # 10KB
)

1000.times do |i|
  msg2 = BasicTestProto2::TestMessage2.new(foo: i)
  msg.optional_msg = msg2  # This causes arena fusion

  msg2 = nil
end
```

This will cause every msg and msg2 to be fused together, but none of the
resulting memory will ever be able to be GC'd because all arenas are
linked to the arena backing the msg, which has a static lifetime. This
will cause unbounded memory growth.

The advised solution to this problem is to use `deep_copy` to copy msg2
before assigning it - which will allow the original msg2 to be
reclaimed.

But this issue is reasonably hard to identify in non-trivial
applications, so the hope is that we'd be able to enable warnings as a
debug step and determine how large our arena chains are getting.