Use the reported byte length for Ruby strings.
We can avoid calling `strlen(3)` on the byte sequence associated with a Ruby string by reading its stored byte size. As a caveat, Ruby strings can contain NULL bytes so there are situations where `strlen(3)` will return a different value than `String#bytesize`. However, that does not matter in this case. The strings are used as keys in a lookup table and will simply fail to match. There's no risk of invalid data because the pointer is owned by the Ruby string.